Renamed RoleHelper to RoleDeterminer.
